refinement for sharing common methods
